requirements.txt
hydra-core
evaluate==0.4.3
absl-py==2.1.0
nltk==3.9.1
rouge_score==0.1.2
tensordict<=0.6.2
trl=0.11.3
accelerate
codetiming
peft
pybind11
pylatexenc
torchdata
wandb
flash_attn